requirements.txt
openai-agents
gradio
fastapi
uvicorn[standard]
python-docx
python-dotenv
httpx
psycopg[binary]
pydantic

